requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
distlib==0.3.1
Django==3.1.7
django-filter==21.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.13.2
filelock==3.0.12
inflection==0.5.1
jsonschema==3.2.0
Pillow==8.1.2
postgres==4.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
PyJWT==2.3.0
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
six==1.15.0
sqlparse==0.4.2
uritemplate==4.1.1
virtualenv==20.4.2
